+++ packages/server/src/queue/implementations/consumers/utilities.ts
@@ -12,15 +12,15 @@ import {
   getSigningKey,
 } from "@iguhealth/jwt";
 
-import { getIssuer } from "../../authN/oidc/constants.js";
-import { WORKER_APP } from "../../authN/oidc/hardcodedClients/worker-app.js";
-import RedisCache from "../../cache/providers/redis.js";
-import { getCertConfig } from "../../certification.js";
-import { createArtifactMemoryDatabase } from "../../fhir-clients/clients/memory/async.js";
-import { createLogger, getRedisClient } from "../../fhir-server/index.js";
-import { IGUHealthServerCTX } from "../../fhir-server/types.js";
-import createStore from "../../storage/index.js";
-import RedisLock from "../../synchronization/redis.lock.js";
+import { getIssuer } from "../../../authN/oidc/constants.js";
+import { WORKER_APP } from "../../../authN/oidc/hardcodedClients/worker-app.js";
+import RedisCache from "../../../cache/providers/redis.js";
+import { getCertConfig } from "../../../certification.js";
+import { createArtifactMemoryDatabase } from "../../../fhir-clients/clients/memory/async.js";
+import { createLogger, getRedisClient } from "../../../fhir-server/index.js";
+import { IGUHealthServerCTX } from "../../../fhir-server/types.js";
+import createStore from "../../../storage/index.js";
+import RedisLock from "../../../synchronization/redis.lock.js";
 
 export type IGUHealthWorkerCTX = Pick<
   IGUHealthServerCTX,
